+++ src/aafc/submodules/modifiers.h
@@ -6,7 +6,7 @@
 #include <aafc.h>
 
 void forceMono(float* input, AAFC_HEADER* h);
-float* resampleAudio(float* input, AAFC_HEADER* header, unsigned long samplerateoverride, float pitch, bool nointerp);
+float* resampleAudio(float* input, AAFC_HEADER* header, unsigned int samplerateoverride, float pitch, bool nointerp);
 float* forceIndependentChannels(float* input, const AAFC_HEADER* h);
 float* normalize(float* input, const AAFC_HEADER* h);
 float* forceInterleaveChannels(float* input, const AAFC_HEADER* h);
