+++ src/aafc/submodules/sfpcm.c
@@ -16,7 +16,7 @@ void* encode_sfpcm(float* ptr, const AAFC_HEADER* h, size_t* audsize) {
             *audsize = h->samplelength;
             unsigned char* const stbs = (unsigned char*)malloc(h->samplelength);
             unsigned char* sptr = stbs;
-            for (unsigned long i = 0; i < h->samplelength; ptr++, sptr++, i++) {
+            for (unsigned int i = 0; i < h->samplelength; ptr++, sptr++, i++) {
                 *sptr = minifloat(CLAMP(*ptr * 127.0f, -128.0f, 127.0f));
             }
             return stbs;
@@ -25,7 +25,7 @@ void* encode_sfpcm(float* ptr, const AAFC_HEADER* h, size_t* audsize) {
             *audsize = h->samplelength * sizeof(short);
             unsigned short* const stbs = (unsigned short*)malloc(*audsize);
             unsigned short* sptr = stbs;
-            for (unsigned long i = 0; i < h->samplelength; ptr++, sptr++, i++) {
+            for (unsigned int i = 0; i < h->samplelength; ptr++, sptr++, i++) {
                 *sptr = halfpercision(*ptr);
             }
             return stbs;
